fmd: Fix the type/size of PCD IOCTL parameters
Copy from user a structure which had uint8_t instead uint16_t,
thus the values after were wrong

[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]

+++ drivers/net/ethernet/freescale/dpa/NetCommSw/src/wrapper/lnxwrp_ioctls_fm_compat.h
@@ -195,7 +195,7 @@ typedef struct ioc_compat_fm_pcd_cc_key_params_t {
 } ioc_compat_fm_pcd_cc_key_params_t;
 
 typedef struct ioc_compat_keys_params_t {
-    uint8_t                                    num_of_keys;
+    uint16_t                                   num_of_keys;
     uint8_t                                    key_size;
     ioc_compat_fm_pcd_cc_key_params_t          key_params[IOC_FM_PCD_MAX_NUM_OF_KEYS]; /**< compat structure*/
     ioc_compat_fm_pcd_cc_next_engine_params_t  cc_next_engine_params_for_miss;         /**< compat structure*/
@@ -217,7 +217,7 @@ typedef struct ioc_compat_fm_pcd_cc_node_modify_key_params_t {
 
 typedef struct ioc_compat_fm_pcd_cc_node_modify_key_and_next_engine_params_t {
     compat_uptr_t                       id;
-    uint8_t                             key_indx;
+    uint16_t                            key_indx;
     uint8_t                             key_size;
     ioc_compat_fm_pcd_cc_key_params_t   key_params;
 } ioc_compat_fm_pcd_cc_node_modify_key_and_next_engine_params_t;
